third_party/blink/web_tests/platform/generic/fast/events/touch/layout-triggers-expected.txt

Check which type of events trigger layout when it's dirty. Events that do a hit-test must update layout, but all other events shouldn't need to.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Forcing initial layout
PASS internals.needsLayoutCount() is 0
Sending touchstart
PASS triggeredLayout is true
Sending touchmove
PASS triggeredLayout is false
Sending touchend
PASS triggeredLayout is false
Sending GestureTapDown
PASS triggeredLayout is true
Sending GestureShowPress
PASS triggeredLayout is true
PASS successfullyParsed is true

TEST COMPLETE
